API Proper Validation and File Organization

+++ pages/api/announcement/[id]/delete.ts
@@ -19,7 +19,7 @@ export default async function editAnnouncement(
     encryption,
   });
 
-  if (token) {
+  if (token?.role && token?.role == "Government") {
     try {
       const deletedAnnouncement = await prisma.announcement.delete({
         where: {
@@ -33,7 +33,7 @@ export default async function editAnnouncement(
       res.status(400).end();
     }
   } else {
-    res.statusMessage = "Unauthorized Access. Please Log In first.";
-    res.status(401).end();
+    res.statusMessage = "Forbidden Access.";
+    res.status(403).end();
   }
 }
